feat(angular-rspack): use development config for serve
