+++ e2e/fixtures/rspack-csr-css/package.json
@@ -9,10 +9,16 @@
   "nx": {
     "targets": {
       "serve-api": {
+        "continuous": true,
         "command": "node ./src/api.mjs",
         "options": {
           "cwd": "e2e/fixtures/rspack-csr-css"
         }
+      },
+      "serve": {
+        "dependsOn": [
+          "serve-api"
+        ]
       }
     }
   }
