+++ packages/angular-rspack/src/lib/config/config-utils/server-config.ts
@@ -1,0 +1,105 @@
+import { Configuration, ContextReplacementPlugin } from '@rspack/core';
+import { resolve } from 'path';
+import {
+  getAllowedHostsConfig,
+  getProxyConfig,
+} from './dev-server-config-utils';
+import { getOptimization } from './optimization-config';
+import { isPackageInstalled } from '../../utils/misc-helpers';
+import { NgRspackPlugin } from '../../plugins/ng-rspack';
+import {
+  I18nOptions,
+  NormalizedAngularRspackPluginOptions,
+} from '../../models';
+
+export async function getServerConfig(
+  root: string,
+  normalizedOptions: NormalizedAngularRspackPluginOptions,
+  i18n: I18nOptions,
+  defaultConfig: Configuration
+): Promise<Configuration> {
+  return {
+    ...defaultConfig,
+    name: 'server',
+    target: 'node',
+    entry: {
+      server: {
+        import: [
+          ...(isPackageInstalled(root, '@angular/platform-server')
+            ? // This import must come before any imports (direct or transitive) that rely on DOM built-ins being
+              // available, such as `@angular/elements`.
+              ['@angular/platform-server/init']
+            : []),
+          ...(i18n.shouldInline ? ['@angular/localize/init'] : []),
+          (normalizedOptions.ssr as { entry: string }).entry,
+        ],
+      },
+    },
+    output: {
+      ...defaultConfig.output,
+      path: normalizedOptions.outputPath.server,
+      filename: '[name].js',
+      chunkFilename: '[name].js',
+      library: { type: 'commonjs' },
+    },
+    devServer: {
+      headers: {
+        'Access-Control-Allow-Origin': '*',
+      },
+      allowedHosts: getAllowedHostsConfig(
+        normalizedOptions.devServer.allowedHosts,
+        normalizedOptions.devServer.disableHostCheck
+      ),
+      client: {
+        webSocketURL: {
+          hostname: normalizedOptions.devServer.host,
+          port: normalizedOptions.devServer.port,
+        },
+        overlay: {
+          errors: true,
+          warnings: false,
+          runtimeErrors: true,
+        },
+        reconnect: true,
+      },
+      host: normalizedOptions.devServer.host,
+      port: normalizedOptions.devServer.port,
+      hot: false,
+      liveReload: true,
+      watchFiles: ['./src/**/*.*', './public/**/*.*'],
+      historyApiFallback: {
+        index: '/index.html',
+        rewrites: [{ from: /^\/$/, to: 'index.html' }],
+      },
+      devMiddleware: {
+        writeToDisk: (file) => !file.includes('.hot-update.'),
+      },
+      server: {
+        options:
+          normalizedOptions.devServer.sslKey &&
+          normalizedOptions.devServer.sslCert
+            ? {
+                key: resolve(root, normalizedOptions.devServer.sslKey),
+                cert: resolve(root, normalizedOptions.devServer.sslCert),
+              }
+            : {},
+        type: normalizedOptions.devServer.ssl ? 'https' : 'http',
+      },
+      proxy: await getProxyConfig(
+        root,
+        normalizedOptions.devServer.proxyConfig
+      ),
+    },
+    externals: normalizedOptions.externalDependencies,
+    optimization: getOptimization(normalizedOptions, 'server'),
+    plugins: [
+      ...(defaultConfig.plugins ?? []),
+      // Fixes Critical dependency: the request of a dependency is an expression
+      new ContextReplacementPlugin(/@?hapi|express[\\/]/),
+      new NgRspackPlugin(normalizedOptions, {
+        i18nOptions: i18n,
+        platform: 'server',
+      }),
+    ],
+  };
+}
