+++ packages/angular-rspack/src/lib/index.ts
@@ -1,2 +1,6 @@
 export * from './config/create-config';
 export * from './plugins/ng-rspack';
+export { configureSourceMap } from './config/config-utils/sourcemap-utils';
+export { getCrossOriginLoading } from './config/config-utils/helpers';
+export { getOptimization } from './config/config-utils/optimization-config';
+export { VENDORS_TEST } from './config/config-utils/optimization-config';
